feat: browse PR comments one by one.
Comment id is an interface as when fetching PR comments it's a string with numbers and letters, when fetching commit comments it's a number.

User exists on Comment as when fetching commit comments the author is called user, so we custom parse it to ensure we can call author everywhere

When we get commit comments we hit the github api directly, which returns snake case json, when we fetch PR details we use the GH cli to wrap a bunch of API calls for us, which returns camel case json, hence the need for a Created field in a comment, to map to ``created_at`.

We want a Set for reviewers to avoid repeating names if they've reviewed multiple times. Go doesn't offer this so we use a map to track which users we've added already.