feat: store history of viewed PRs
Now we can let the user know when there's new comments so they don't have to read everything again to know if that's the case.

Interfaces and mocks galore! Now we can start testing the pr_action as we can easily stub everything, including printing.

To enable the tests we've moved the `R` call up to root, otherwise we'd have an infinite loop in our test