chore(refactor): move runcommand into it's own client + add http client
When we add review comments we'll need to fetch the auth token which requires running the gh cli so if we move the runcommand method out to it's own thing, we can access it from the root command and get to the auth token.

Http client will allow us to make requests to any url (e.g. Github graphql). Auth token is stored in the client so we don't have to keep fetching it when we want it (e.g. when replying or resolving a thread)