feat: include review comments in list
This allows people to read comments posted as part of a review. They aren't available via the command line and need graphql. We could use graphql for all of it but then we'd have to do a bunch of requests, using the commandline where we can hides some of that for us