chore(refactor): use graphql for all queries to github
This is simpler than messing with the GH CLI and makes requests consistent. We don't request an id on reviews as we don't use it, some data we only need for verbose so to improve request speed don't include it.

Now we no longer need the Flatten helper or getCommitComments as it's part of the single graphql request for all PR information

Nor do we need custom json parsing to handle the different created/submitted user/author formats between the cli and graphql