feat: print file details if the comment is on a new line/file
It's useful to know what the comment is referring to, in the case of comments on the PR itself we default to "main thread" to indicate that.

Used a separate method for printing the actual contents to allow us to call it from an 'expand' command and bypass the hiding the body if it's resolved/outdated as expand is specifically for viewing those comments.

Reset the last full path on expand to force it to show the file info, otherwise the last full path is the current item but the dev didn't see those details as they were hidden by the outdated/resolved logic