feat: make pr number optional
By using the current branch we can filter Github for the open PR and use that, saving the dev from having to find it and type it.
We still want them to be able to browse other PRs so move the argument validation into pr_action. We could have multiple init functions instead, or pass a negative number for no pr number specified, or use function options. Each has pros and cons, multiple init could be more confusing and given the need to fetch repo info is more involved, negative number is a bit cryptic and function options hide what's going on and add more complexity. Moving argument validation into the action means we can test it (yay!) and have a single init method so is hopefully easier to understand.

Re-used the gh cli git client so we don't have to parse the current branch ourselves, added an interface to make it easily testable.

You shouldn't ever have more than 1 PR for a branch, that implies something's broken on Github's end but in case they ever implement that or things have gone wonky, protect against it