feat: allow checking owned PRs for new comments
Using the `peruse pr check` command you can list all PRs that have new comments since you last looks, also `-n` will show a notification so it can be used in a scheduled job

make `Print` and `Println` of Output return an error so notify can implement the same interface and let the caller know it failed. Now the "whether to notify or print" logic lives in the run of the command which feels nicer and easier to manage than burying it in `CheckForNewComments`. We can ignore that error in pr_action as it's always nil for this scenario of printing.

Used a mega Github search query to get all owned PRs and comment count so we can do one request for all the info

Different functions for notifying for different OS's so we don't have to detect the OS at runtime. Used Go's build tags so it adds the right one to the binary for each platform

Limited to 20 PRs to stop Github getting upset with us