feat: copy comment to clipboard
With the command 'x' yo can now copy the current comment to the clipboard.

pbcopy and clip are built into MacOS/Windows but Linux doesn't have anything by default. We could use some C bindings but it's a lot simpler to ask people to install a package. Native implementations can come later!

Using a variable for WriteTo so it's overwriteable in the test, not great but given we use different implementations for different os's seems the most straight forward. This can be replicated for Notify too.

X was chosen as 'c' is comment, which is probably more common and 'x' is used for cut so is sort of intuitive