chore(tests): refactor to allow testing prompt
By supplying the prompt we can mock it and test it to check the various commands do what they should.

Similarly, moved the switch block into it's own method so we can test outside of the while loop. Just need to sort out the os.exit call and then all branches are tested